requirements.txt
fastapi
uvicorn
pandas
openpyxl
requests
msal
google-generativeai
pydantic
python-dotenv
cachetools
